requirements.txt
streamlit
fastapi
uvicorn
python-docx
scikit-learn
pandas
nltk 